[INTERNAL][FIX] ComponentContainer: don't ignore component's parent
If `propagetModel` is not set to true, the ComponentContainer still
propagated empty model and binding context sets as well as any
registered propagation listeners.

But this conflicts with a parent of the component, if any exists.
Therefore, the reduced propagation must only happen if the component
has no parent.

SNOW: DINC0569532
SNOW: DINC0568593
SNOW: DINC0569239
Change-Id: Iabd578f90f601f802a2a74657942d4c6687a8055