[INTERNAL][FIX] sap.ui.fl: fix timing issue in VariantModel
There was a timing issue in the VariantModel between the initial call
to the model from the control (registerToModel) and a subsequent call
to activateVariant.
To resolve this the internal variantSwitchPromise, which previously
was responsible for all vm controls, is split so that each vm control
has its own variantSwitchPromise.

SNOW: CS20250009808153
CR-Id: 002075125800002607252025
Change-Id: I949d282e13ff4a8716b26bd25b33655ca2505e94
(cherry picked from commit de06ee8ed70f425517d967c052e9d135b1296e05)