[INTERNAL][FIX] sap.ui.fl: fix timing issue in VariantModel
There was a timing issue in the VariantModel between the initial call
to the model from the control (registerToModel) and a subsequent call
to activateVariant.
To resolve this the internal variantSwitchPromise, which previously
was responsible for all vm controls, is split so that each vm control
has its own variantSwitchPromise.

SNOW: CS20250009808153
Change-Id: I949d282e13ff4a8716b26bd25b33655ca2505e94