[INTERNAL][FIX] ui.Table: moved column related announcements to column header
The "fixed column" announcement was added on focus on header and cell level and not cleaned up which caused double announcement in some cases. Now it is only added on header level during rendering.

The "required" announcement was only added on focus on header level, and in some cases not announced at all. Now also adding this association during rendering.

SNOW: DINC0553938
Change-Id: I4449644291c8096155ce3501af531a857325055f