[FIX] sap.m.OverflowToolbar: fix scrolling down the page
Problem: When navigating with down arrow key and reaching the last
Control of the OverflowToolbar, page does not get scrolled down.

Solution: The problem happens because overflow buttons (...) is
considered "interactive" and is added to the interactive controls array,
even when it is not currently visible. As a result, when navigating with
down arrow key from the last Control, OFT is trying to move the focus on
the overflow button (...), which is a wrong behavior.
Now we do not add the overflow button to the interactive controls array
when it is not visible.

SNOW: DINC0558087
Change-Id: I466c8956e01fe4b0c636eae9112356b4c77913e9