[FIX] sap.uxap.BlockBase: fix table cutting
Problem: Table gets cut and no scrollbar on the ObjectPageLayout is
available when screen size is too small and table has already reached
its minimum height (depending on the minRowCount). This only happens
when the SubSection has "sapUxAPObjectPageSubSectionFitContainer" class
and its content is placed in BlockBase.

Solution: ObjectPageSubSection cannot properly handle this situation
because BlockBase has overflow-y: hidden, which cuts the table and does
not allow the ObjectPageSubSection to know that its content now does not
fit (scrollHeight is bigger than height). It seems this CSS style was
added some time ago for an edge case situation, but it may be redundant
as multiple changes has been made on ObjectPageSubSection with
fitContent class during this time, which improved the handling of such
situations.

SNOW: DINC0507572
Change-Id: Ie8bb656ec3eb67135472a9f406affbe9edb158b6