[FEATURE] Table: Add ODataV4-specific plugins for selection and tree support
The ODataV4MultiSelection and ODataV4SingleSelection plugins integrate
the table with the selection built into the
sap.ui.model.odata.v4.ODataListBinding. The selection is visualized in
the table and can be changed interactively by the user. To work with the
selection programmatically, for example read or change the selection,
the bindings API can be used.

The ODataV4Aggregation and ODataV4Hierarchy plugins integrate the table
with the hierarchy and data aggregation capabilities of the
sap.ui.model.odata.v4.ODataListBinding. The data structure is visualized
in the table and the user can interactively expand and collapse rows.

Change-Id: Iea28b19c709920dcd913e79f9cd41d278d36be5d
Jira: CPOUIFTEAMB-2418