[FIX] sap.ui.model.odata.v2.ODataAnnotations: Propagate 'withCredentials' flag when loading annotations
The problem:
When the ODataAnnotations are loaded separately from the (via external
URL) ODataModel, the 'withCredentials' flag was not being passed to the
Ajax request. The metadata reques propagate this flag i.e. if the
annotations would be loaded with the metadata the flag would have been
passed.

PS1: Do it!
PS2: Review comments; Add test that withCredentials is propagated from
	Models constructor
PS3-5: Review comments

SNOW: CS20250009642388
Change-Id: Idf6d2bc367f7b3a25584b8f47ef2b20e3be25a95