[INTERNAL][FIX] sap.ui.dt: DesignTimeMetadata - fix bundle handling (reinstatement)
When retrieving the resource bundle from control libraries, we were
trying to load the bundle as if it were a library - which caused an
extra request for a manifest file that does not exist.

This change uses the proper method to create the bundle with its path.

This is a reinstatement of a previous fix that was reverted because the
core did't support cashing of expensive operations in the ResourceBundle

JIRA: TEAM42-415
Change-Id: I88dc6a0069fcbb299ee0f46d14dfec007a2ade48