[INTERNAL] ColumnUtils: Fix reordering of first column in tree mode
Preparation for CPOUIFTEAMB-2418

In a tree table, the first visible column must not be movable, and it must not be possible to move another column before the first visible column. This logic is now applied for any table that is in tree mode.

Fixes a bug if there are invisible columns in the beginning. It was possible to move the first visible column in of a table in tree mode. This is now correctly prevented.

Change-Id: Ia1514e586b7858d34ea05af630ba65e3b781659b
SNOW: DINC0576479