[INTERNAL] sap.f.FlexibleColumnLayout: Update routing logic in samples
Problem: Samples of FlexibleColumnLayout work with async views loading.
However, in this case, onBeforeRouteMatch event cannot be used in the
controllers because with async views loading, we cannot rely on the
sequence of events that are fired during the loading phase. More
secifically, in this case, beforeRouteMatch event is fired BEFORE the
controller is even initialized.

Solution: Attach the beforeRouteMatch event handler in the Component's
init method. This ensures that the handler is registered before any
routing events are fired.

JIRA: BGSOFUIPIRIN-6833
Change-Id: Ica74e5a943f61b40a21516f87fd427e5e11e14c5