[INTERNAL] sap.ui.fl: fix index-relevant condensing
If all controls in an aggregation are touched by a change, they just get sorted
by the target index. This does not work for Destroys, as they don't have
a target index. Instead they get the index -1 assigned.
This way they get sorted to the front, syncing the number of items in the
aggregation between the target UI and the remaining changes.

SNOW: CS20250010290857
Change-Id: Ibb0848e14621a32105295197cf65837030ff252d