[FIX] sap.m.Menu: fixed menu opening when items are bound to OData V4 model
Issue: After the menu refactoring (Gerrit change #6460577), when the
items are bound to a OData v4 model, menu cannot be opened as after
the data receiving there is a re-rendering which causes loss of focus
of the menu popover, and it is auto-closed.

Solution: now if the items are bound to OData v 4 model, the menu's
popover is being opened when the data is fully received, and is not
auto-closed.

Test is manual, as the issue is reproducible only with OData v4.
Snippix for test: #712934

SNOW: DINC0574936
SNOW: DINC0576340
Change-Id: Id9d830f9108a014f2b5e58df4805844487cbfdc1