[FIX] sap.m.SinglePlanningCalendarMonthGrid: fixed Month view layout (320-340px)
Issue: on desktop devices where the week number is displayed on Month
view, there was displacement of the days of week header and the dates
below because of the min-width set for day cells.

Solution: now the min-width is removed for smaller resolutions, and
the issue is fixed.

SNOW: DINC0172416
Change-Id: Iafd6225050a0bd8fb49bc9c9c662f04ae2346f03