[FIX] PropertyHelperUtil: check UI5 2.0 availability by deprecations
ISSUE: Previously, we enabled propertyInfo validation in Ui5 2.0. For this, we loaded the VersionInfo and checked whether it is 2.0. This does not always work. In some cases, VersionInfo.load to to errors, resulting in missing bindingContexts for MDC Table.

FIX: Instead of checking the version, we now check for the existence of deprecated functions. Specifically Core.attachInit. If available, Ui5 version is < 2.0.

SNOW: DINC0578994
CR-Id: 002075125800002983572025
Change-Id: Ib35bf8659fb3e2642c44050f6c89d5055b831418
(cherry picked from commit 2b3690e7e91869af60d9314a305580202510eb03)