[INTERNAL] odata.v4.TableDelegate: Remove changed path handling on rebind
The logic to force a rebind when changing the binding path did not work with named models. If the path in the binding info is, for example, "myModel>/myPath", Binding#getPath returns "myPath".

A more sophisticated check could be added. But changing the path usually means changing the entity the table is bound to, which is basically a metadata change. The PropertyInfos are static. If a delegate needs to support this scenario, it needs to be handled in the subclass (for example call Delegate#rebind inside Delegate#updateBinding if the path changed).

Change-Id: I69e2d4018866c07d416298121a2adf3dbf725f69